# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Connect to MongoDB API.",
        "title": "Connect to MongoDB API",
        "contact": {},
        "version": "1.0"
    },
    "host": "modul-go.onrender.com",
    "basePath": "/",
    "paths": {
        "/analyze": {
            "get": {
                "description": "Analyzes user activities and returns a report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Analyze user activities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User login",
                        "name": "login",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/connect-to-mongodb_internal_analysis.AnalysisReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/import": {
            "post": {
                "description": "Imports users from a JSON payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Import users from JSON",
                "parameters": [
                    {
                        "description": "JSON payload containing user objects",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/report": {
            "get": {
                "description": "Generates a JSON report of all user activities",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Generate JSON report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/connect-to-mongodb_internal_analysis.AnalysisReport"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "delete": {
                "description": "Deletes all users from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Delete all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{login}": {
            "delete": {
                "description": "Deletes a user by their login from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Delete user by login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User login",
                        "name": "login",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ws/logs": {
            "get": {
                "description": "Logs HTTP requests for monitoring and debugging",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logging"
                ],
                "summary": "Log HTTP requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "connect-to-mongodb_internal_analysis.Activity": {
            "type": "object",
            "properties": {
                "activity_description": {
                    "type": "string"
                },
                "activity_name": {
                    "type": "string"
                },
                "date": {},
                "id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "time": {}
            }
        },
        "connect-to-mongodb_internal_analysis.ActivityLimit": {
            "type": "object",
            "properties": {
                "activity_name": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                }
            }
        },
        "connect-to-mongodb_internal_analysis.AnalysisReport": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/connect-to-mongodb_internal_analysis.Activity"
                    }
                },
                "activities_summary_minutes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "daily_limits": {
                    "$ref": "#/definitions/connect-to-mongodb_internal_analysis.DailyLimits"
                },
                "date": {
                    "type": "string"
                },
                "exceeded_limits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "global_limit": {
                    "type": "string"
                },
                "global_limit_exceeded": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/connect-to-mongodb_internal_analysis.Notification"
                    }
                },
                "total_duration_minutes": {
                    "type": "integer"
                }
            }
        },
        "connect-to-mongodb_internal_analysis.DailyLimits": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/connect-to-mongodb_internal_analysis.ActivityLimit"
                    }
                },
                "global_limit": {
                    "type": "integer"
                }
            }
        },
        "connect-to-mongodb_internal_analysis.Notification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "read": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}